nota-fiscal-python/requirements.txt
Flask==3.0.3
psycopg2-binary==2.9.9
reportlab==4.2.2
python-dotenv==1.0.1
